requirements.txt
fastapi[standard]
gunicorn
requests
python-dotenv